feat: Auto-attach current theme to prompt after 1st user message
